Fixed precision issues by using BigDecimal instead of float for currency amounts.

+++ src/main/java/com/mediamath/terminalone/models/AudienceSegment.java
@@ -16,6 +16,7 @@
 
 package com.mediamath.terminalone.models;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
@@ -122,7 +123,7 @@ public class AudienceSegment implements T1Entity {
 		this.retail_cpm = retail_cpm;
 	}
 
-	public void setRetailCpm(float retail_cpm) {
+	public void setRetailCpm(BigDecimal retail_cpm) {
 		Currency curr = new Currency();
 		curr.setValue(retail_cpm);
 		this.retail_cpm.add(curr);
@@ -136,7 +137,7 @@ public class AudienceSegment implements T1Entity {
 		this.wholesale_cpm = wholesale_cpm;
 	}
 
-	public void setWholesaleCpm(float wholesale_cpm) {
+	public void setWholesaleCpm(BigDecimal wholesale_cpm) {
 		Currency curr = new Currency();
 		curr.setValue(wholesale_cpm);
 		this.wholesale_cpm.add(curr);
